+++ src/core.js
@@ -892,13 +892,14 @@ goog.require('olcs.core.OlLayerPrimitive');
    * @param {!ol.geom.Geometry} geometry Ol3 geometry.
    * @param {!ol.proj.ProjectionLike} projection
    * @param {!ol.style.Style} olStyle
+   * @param {!Cesium.BillboardCollection} billboards
    * @param {function(!Cesium.Billboard)=} opt_newBillboardCallback Called when
    * the new billboard is added.
-   * @return {!Cesium.Primitive} primitives
+   * @return {Cesium.Primitive} primitives
    * @api
    */
   olcs.core.olMultiGeometryToCesium = function(geometry, projection,
-      olStyle, opt_newBillboardCallback) {
+      olStyle, billboards, opt_newBillboardCallback) {
     // Do not reproject to 4326 now because it will be done later.
 
     // FIXME: would be better to combine all child geometries in one primitive
@@ -917,7 +918,6 @@ goog.require('olcs.core.OlLayerPrimitive');
         geometry = /** @type {!ol.geom.MultiPoint} */ (geometry);
         subgeos = geometry.getPoints();
         var fn = olcs.core.olPointGeometryToCesium;
-        var billboards = new Cesium.BillboardCollection();
         if (olStyle.getText()) {
           var primitives = new Cesium.PrimitiveCollection();
           goog.array.forEach(subgeos, function(geometry) {
@@ -935,7 +935,7 @@ goog.require('olcs.core.OlLayerPrimitive');
             fn(geometry, projection, olStyle, billboards,
                 opt_newBillboardCallback);
           });
-          return billboards;
+          return null;
         }
       case 'MultiLineString':
         geometry = /** @type {!ol.geom.MultiLineString} */ (geometry);
@@ -1206,8 +1206,14 @@ goog.require('olcs.core.OlLayerPrimitive');
       case 'MultiPoint':
       case 'MultiLineString':
       case 'MultiPolygon':
-        return id(olcs.core.olMultiGeometryToCesium(geom, proj, style,
-            newBillboardAddedCallback));
+        var result = olcs.core.olMultiGeometryToCesium(geom, proj, style,
+            context.billboards, newBillboardAddedCallback);
+        if (!result) {
+          // no wrapping primitive
+          return null;
+        } else {
+          return id(result);
+        }
       case 'LinearRing':
         throw new Error('LinearRing should only be part of polygon.');
       default:
@@ -1223,11 +1229,12 @@ goog.require('olcs.core.OlLayerPrimitive');
    * @param {!ol.layer.Vector} olLayer
    * @param {!ol.View} olView
    * @param {!Object.<number, !Cesium.Primitive>} featurePrimitiveMap
+   * @param {!Cesium.Scene} scene
    * @return {!olcs.core.OlLayerPrimitive}
    * @api
    */
   olcs.core.olVectorLayerToCesium = function(olLayer, olView,
-      featurePrimitiveMap) {
+      featurePrimitiveMap, scene) {
     var vectorLayer = olLayer;
     var features = vectorLayer.getSource().getFeatures();
     var proj = olView.getProjection();
@@ -1239,7 +1246,7 @@ goog.require('olcs.core.OlLayerPrimitive');
       // are defined
       throw new Error('View not ready');
     }
-    var allPrimitives = new olcs.core.OlLayerPrimitive(proj);
+    var allPrimitives = new olcs.core.OlLayerPrimitive(proj, scene);
     var context = allPrimitives.context;
     for (var i = 0; i < features.length; ++i) {
       var feature = features[i];