+++ src/core/ollayerprimitive.js
@@ -6,12 +6,13 @@ goog.provide('olcs.core.OlLayerPrimitive');
  * Result of the conversion of an OpenLayers layer to Cesium.
  * @constructor
  * @param {!(ol.proj.Projection|string)} layerProjection
+ * @param {!Cesium.Scene} scene
  * @extends {Cesium.PrimitiveCollection}
  */
-olcs.core.OlLayerPrimitive = function(layerProjection) {
+olcs.core.OlLayerPrimitive = function(layerProjection, scene) {
   goog.base(this);
 
-  var billboards = new Cesium.BillboardCollection();
+  var billboards = new Cesium.BillboardCollection({scene: scene});
   var primitives = new Cesium.PrimitiveCollection();
 
   /**